+++ deployment/run-unit-tests.sh
@@ -2,5 +2,15 @@
 
 set -e
 
-cd ../source/image-handler
+current_dir=$PWD
+source_dir=$current_dir/../source
+
+cd $source_dir/constructs
+npm install
+npm test
+
+cd $source_dir/image-handler
+npm test
+
+cd $source_dir/custom-resource
 npm test