+++ source/image-handler/image-request.js
@@ -1,32 +1,78 @@
-/*********************************************************************************************************************
- *  Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.                                           *
- *                                                                                                                    *
- *  Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance    *
- *  with the License. A copy of the License is located at                                                             *
- *                                                                                                                    *
- *      http://www.apache.org/licenses/LICENSE-2.0                                                                    *
- *                                                                                                                    *
- *  or in the 'license' file accompanying this file. This file is distributed on an 'AS IS' BASIS, WITHOUT WARRANTIES *
- *  OR CONDITIONS OF ANY KIND, express or implied. See the License for the specific language governing permissions    *
- *  and limitations under the License.                                                                                *
- *********************************************************************************************************************/
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
 
 const ThumborMapping = require('./thumbor-mapping');
 
 class ImageRequest {
+    constructor(s3, secretsManager) {
+        this.s3 = s3;
+        this.secretsManager = secretsManager;
+    }
 
     /**
      * Initializer function for creating a new image request, used by the image
      * handler to perform image modifications.
-     * @param {Object} event - Lambda request body.
+     * @param {object} event - Lambda request body.
      */
     async setup(event) {
         try {
+            // Checks signature enabled
+            if (process.env.ENABLE_SIGNATURE === 'Yes') {
+                const crypto = require('crypto');
+                const { path, queryStringParameters } = event;
+                if (!queryStringParameters || !queryStringParameters.signature) {
+                    throw {
+                        status: 400,
+                        message: 'Query-string requires the signature parameter.',
+                        code: 'AuthorizationQueryParametersError'
+                    };
+                }
+
+                const { signature } = queryStringParameters;
+                try {
+                    const response = await this.secretsManager.getSecretValue({ SecretId: process.env.SECRETS_MANAGER }).promise();
+                    const secretString = JSON.parse(response.SecretString);
+                    const hash = crypto.createHmac('sha256', secretString[process.env.SECRET_KEY]).update(path).digest('hex');
+
+                    // Signature should be made with the full path.
+                    if (signature !== hash) {
+                        throw {
+                            status: 403,
+                            message: 'Signature does not match.',
+                            code: 'SignatureDoesNotMatch'
+                        };
+                    }
+                } catch (error) {
+                    if (error.code === 'SignatureDoesNotMatch') {
+                        throw error;
+                    }
+
+                    console.error('Error occurred while checking signature.', error);
+                    throw {
+                        status: 500,
+                        message: 'Signature validation failed.',
+                        code: 'SignatureValidationFailure'
+                    };
+                }
+            }
+
             this.requestType = this.parseRequestType(event);
             this.bucket = this.parseImageBucket(event, this.requestType);
             this.key = this.parseImageKey(event, this.requestType);
             this.edits = this.parseImageEdits(event, this.requestType);
             this.originalImage = await this.getOriginalImage(this.bucket, this.key);
+            this.headers = this.parseImageHeaders(event, this.requestType);
+
+            if (!this.headers) {
+                delete this.headers;
+            }
+
+            // If the original image is SVG file and it has any edits but no output format, change the format to WebP.
+            if (this.ContentType === 'image/svg+xml'
+                && this.edits && Object.keys(this.edits).length > 0
+                && !this.edits.toFormat) {
+                this.outputFormat = 'png'
+            }
 
             /* Decide the output format of the image.
              * 1) If the format is provided, the output format is the provided format.
@@ -56,24 +102,26 @@ class ImageRequest {
                 }
             }
 
-            return Promise.resolve(this);
+            delete this.s3;
+            delete this.secretsManager;
+
+            return this;
         } catch (err) {
-            return Promise.reject(err);
+            console.error(err);
+            throw err;
         }
     }
 
     /**
      * Gets the original image from an Amazon S3 bucket.
-     * @param {String} bucket - The name of the bucket containing the image.
-     * @param {String} key - The key name corresponding to the image.
+     * @param {string} bucket - The name of the bucket containing the image.
+     * @param {string} key - The key name corresponding to the image.
      * @return {Promise} - The original image or an error.
      */
     async getOriginalImage(bucket, key) {
-        const S3 = require('aws-sdk/clients/s3');
-        const s3 = new S3();
         const imageLocation = { Bucket: bucket, Key: key };
         try {
-            const originalImage = await s3.getObject(imageLocation).promise();
+            const originalImage = await this.s3.getObject(imageLocation).promise();
 
             if (originalImage.ContentType) {
                 this.ContentType = originalImage.ContentType;
@@ -95,28 +143,28 @@ class ImageRequest {
                 this.CacheControl = "max-age=31536000,public";
             }
 
-            return Promise.resolve(originalImage.Body);
+            return originalImage.Body;
         } catch(err) {
-            return Promise.reject({
+            throw {
                 status: ('NoSuchKey' === err.code) ? 404 : 500,
                 code: err.code,
                 message: err.message
-            });
+            };
         }
     }
 
     /**
      * Parses the name of the appropriate Amazon S3 bucket to source the
      * original image from.
-     * @param {String} event - Lambda request body.
-     * @param {String} requestType - Image handler request type.
+     * @param {string} event - Lambda request body.
+     * @param {string} requestType - Image handler request type.
      */
     parseImageBucket(event, requestType) {
         if (requestType === "Default") {
             // Decode the image request
             const decoded = this.decodeRequest(event);
             if (decoded.bucket !== undefined) {
-                // Check the provided bucket against the whitelist
+                // Check the provided bucket against the allowed list
                 const sourceBuckets = this.getAllowedSourceBuckets();
                 if (sourceBuckets.includes(decoded.bucket) || decoded.bucket.match(new RegExp('^' + sourceBuckets[0] + '$'))) {
                     return decoded.bucket;
@@ -147,8 +195,8 @@ class ImageRequest {
 
     /**
      * Parses the edits to be made to the original image.
-     * @param {String} event - Lambda request body.
-     * @param {String} requestType - Image handler request type.
+     * @param {string} event - Lambda request body.
+     * @param {string} requestType - Image handler request type.
      */
     parseImageEdits(event, requestType) {
         if (requestType === "Default") {
@@ -186,7 +234,23 @@ class ImageRequest {
         }
 
         if (requestType === "Thumbor" || requestType === "Custom") {
-            return decodeURIComponent(event["path"].replace(/\d+x\d+\/|filters[:-][^/;]+|\/fit-in\/+|^\/+/g,'').replace(/^\/+/,''));
+            let { path } = event;
+
+            if (requestType === "Custom") {
+                const matchPattern = process.env.REWRITE_MATCH_PATTERN;
+                const substitution = process.env.REWRITE_SUBSTITUTION;
+
+                if (typeof(matchPattern) === 'string') {
+                    const patternStrings = matchPattern.split('/');
+                    const flags = patternStrings.pop();
+                    const parsedPatternString = matchPattern.slice(1, matchPattern.length - 1 - flags.length);
+                    const regExp = new RegExp(parsedPatternString, flags);
+                    path = path.replace(regExp, substitution);
+                } else {
+                    path = path.replace(matchPattern, substitution);
+                }
+            }
+            return decodeURIComponent(path.replace(/\/(\d+x\d+)\/|filters:[^\)]+|\/fit-in+|^\/+/g, '').replace(/\)/g, '').replace(/^\/+/, ''));
         }
 
         // Return an error for all other conditions
@@ -202,21 +266,20 @@ class ImageRequest {
      * prefix to the image request. Categorizes a request as either "image"
      * (uses the Sharp library), "thumbor" (uses Thumbor mapping), or "custom"
      * (uses the rewrite function).
-     * @param {Object} event - Lambda request body.
+     * @param {object} event - Lambda request body.
     */
     parseRequestType(event) {
         const path = event["path"];
-        // ----
         const matchDefault = new RegExp(/^(\/?)([0-9a-zA-Z+\/]{4})*(([0-9a-zA-Z+\/]{2}==)|([0-9a-zA-Z+\/]{3}=))?$/);
-        const matchThumbor = new RegExp(/^(\/?)((fit-in)?|(filters:.+\(.?\))?|(unsafe)?).*(.+jpg|.+png|.+webp|.+tiff|.+jpeg)$/i);
-        const matchCustom = new RegExp(/(\/?)(.*)(jpg|png|webp|tiff|jpeg)/i);
+        const matchThumbor = new RegExp(/^(\/?)((fit-in)?|(filters:.+\(.?\))?|(unsafe)?).*(.+jpg|.+png|.+webp|.+tiff|.+jpeg|.+svg)$/i);
+        const matchCustom = new RegExp(/(\/?)(.*)(jpg|png|webp|tiff|jpeg|svg)/i);
         const definedEnvironmentVariables = (
             (process.env.REWRITE_MATCH_PATTERN !== "") &&
             (process.env.REWRITE_SUBSTITUTION !== "") &&
             (process.env.REWRITE_MATCH_PATTERN !== undefined) &&
             (process.env.REWRITE_SUBSTITUTION !== undefined)
         );
-        // ----
+
         if (matchDefault.test(path)) {  // use sharp
             return 'Default';
         } else if (matchCustom.test(path) && definedEnvironmentVariables) {  // use rewrite function then thumbor mappings
@@ -227,21 +290,37 @@ class ImageRequest {
             throw {
                 status: 400,
                 code: 'RequestTypeError',
-                message: 'The type of request you are making could not be processed. Please ensure that your original image is of a supported file type (jpg, png, tiff, webp) and that your image request is provided in the correct syntax. Refer to the documentation for additional guidance on forming image requests.'
+                message: 'The type of request you are making could not be processed. Please ensure that your original image is of a supported file type (jpg, png, tiff, webp, svg) and that your image request is provided in the correct syntax. Refer to the documentation for additional guidance on forming image requests.'
             };
         }
     }
 
     /**
+     * Parses the headers to be sent with the response.
+     * @param {object} event - Lambda request body.
+     * @param {string} requestType - Image handler request type.
+     * @return {object} Custom headers
+     */
+    parseImageHeaders(event, requestType) {
+        if (requestType === 'Default') {
+            const decoded = this.decodeRequest(event);
+            if (decoded.headers) {
+                return decoded.headers;
+            }
+        }
+
+        return undefined;
+    }
+
+    /**
      * Decodes the base64-encoded image request path associated with default
      * image requests. Provides error handling for invalid or undefined path values.
-     * @param {Object} event - The proxied request object.
+     * @param {object} event - The proxied request object.
      */
     decodeRequest(event) {
         const path = event["path"];
         if (path !== undefined) {
-            const splitPath = path.split("/");
-            const encoded = splitPath[splitPath.length - 1];
+            const encoded = path.charAt(0) === '/' ? path.slice(1) : path;
             const toBuffer = Buffer.from(encoded, 'base64');
             try {
                 // To support European characters, 'ascii' was removed.
@@ -288,7 +367,7 @@ class ImageRequest {
     */
     getOutputFormat(event) {
         const autoWebP = process.env.AUTO_WEBP;
-        if (autoWebP && event.headers.Accept && event.headers.Accept.includes('image/webp')) {
+        if (autoWebP === 'Yes' && event.headers.Accept && event.headers.Accept.includes('image/webp')) {
             return 'webp';
         } else if (this.requestType === 'Default') {
             const decoded = this.decodeRequest(event);