+++ CHANGELOG.md
@@ -4,6 +4,61 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.1.0] - 2020-11-19
+### ⚠ BREAKING CHANGES
+- **Image URL Signature**: When image URL signature is enabled, all URLs including existing URLs should have `signature` query parameter.
+
+### Added
+- Image URL signature: [#111](https://github.com/awslabs/serverless-image-handler/issues/111), [#203](https://github.com/awslabs/serverless-image-handler/issues/203), [#221](https://github.com/awslabs/serverless-image-handler/issues/221), [#227](https://github.com/awslabs/serverless-image-handler/pull/227)
+- AWS Lambda `413` error handling. When the response payload is bigger than 6MB, it throws `TooLargeImageException`: [#35](https://github.com/awslabs/serverless-image-handler/issues/35), [#97](https://github.com/awslabs/serverless-image-handler/issues/97), [#193](https://github.com/awslabs/serverless-image-handler/issues/193), [#204](https://github.com/awslabs/serverless-image-handler/issues/204)
+- Default fallback image: [#137](https://github.com/awslabs/serverless-image-handler/issues/137)
+- Unit tests for custom resource: `100%` coverage
+- Add `SVG` support. When any edits are used, the output would be automatically `PNG` unless the output format is specified: [#31](https://github.com/awslabs/serverless-image-handler/issues/31), [#234](https://github.com/awslabs/serverless-image-handler/issues/234)
+- Custom headers: [#182](https://github.com/awslabs/serverless-image-handler/pull/182)
+- Enabling ALB Support : [#201](https://github.com/awslabs/serverless-image-handler/pull/201)
+
+### Fixed
+- Thumbor paths broken if they include "-" and "100x100": [#208](https://github.com/awslabs/serverless-image-handler/issues/208)
+- Rewrite doesn't seem to be working: [#121](https://github.com/awslabs/serverless-image-handler/issues/121)
+- Correct EXIF: [#197](https://github.com/awslabs/serverless-image-handler/issues/197), [#220](https://github.com/awslabs/serverless-image-handler/issues/220), [#235](https://github.com/awslabs/serverless-image-handler/issues/235), [#236](https://github.com/awslabs/serverless-image-handler/issues/236), [#240](https://github.com/awslabs/serverless-image-handler/issues/240)
+- Sub folder support in Thumbor `watermark` filter: [#231](https://github.com/awslabs/serverless-image-handler/issues/231)
+
+### Changed
+- AWS CDK and AWS Solutions Constructs version (from 1.57.0 to 1.64.1)
+- sharp base version (from 0.25.4 to 0.26.1)
+- Refactors the custom resource Lambda source code
+- Migrate unit tests to use `jest`
+- Move all `aws-sdk` in `ImageHandler` Labmda function to `index.js` for the best practice
+- Enhance the default error message not to show empty JSON: [#206](https://github.com/awslabs/serverless-image-handler/issues/206)
+
+### Removed
+- Remove `manifest-generator`
+
+## [5.0.0] - 2020-08-31
+### Added
+- AWS CDK and AWS Solutions Constructs to create AWS CloudFormation template
+
+### Fixed
+- Auto WebP does not work properly: [#195](https://github.com/awslabs/serverless-image-handler/pull/195), [#200](https://github.com/awslabs/serverless-image-handler/issues/200), [#205](https://github.com/awslabs/serverless-image-handler/issues/205)
+- A bug where base64 encoding containing slash: [#194](https://github.com/awslabs/serverless-image-handler/pull/194)
+- Thumbor issues:
+  - `0` size support: [#183](https://github.com/awslabs/serverless-image-handler/issues/183)
+  - `convolution` filter does not work: [#187](https://github.com/awslabs/serverless-image-handler/issues/187)
+  - `fill` filter does not work: [#190](https://github.com/awslabs/serverless-image-handler/issues/190)
+- __Note that__ duplicated features has been merged gracefully.
+
+### Removed
+- AWS CloudFormation template: `serverless-image-handler.template`
+
+### Changed
+- sharp base version (from 0.23.4 to 0.25.4)
+- Remove `Promise` to return since `async` functions return promises: [#189](https://github.com/awslabs/serverless-image-handler/issues/189)
+- Unit test statement coverage improvement:
+  - `image-handler.js`: `79.05%` to `100%`
+  - `image-request.js`: `93.58%` to `100%`
+  - `thumbor-mapping.js`: `99.29%` to `100%`
+  - `overall`: `91.55%` to `100%`
+
 ## [4.2] - 2020-02-06
 ### Added
 - Honor outputFormat Parameter from the pull request [#117](https://github.com/awslabs/serverless-image-handler/pull/117)
@@ -16,7 +71,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - fix: DeprecationWarning: Buffer() is deprecated from the pull request [#174](https://github.com/awslabs/serverless-image-handler/pull/174)
 - Add hex color support for Thumbor ```filters:background_color``` and ```filters:fill``` [#154](https://github.com/awslabs/serverless-image-handler/issues/154)
 - Add format and watermark support for Thumbor [#109](https://github.com/awslabs/serverless-image-handler/issues/109), [#131](https://github.com/awslabs/serverless-image-handler/issues/131), [#109](https://github.com/awslabs/serverless-image-handler/issues/142)
-* __Note that__ duplicated features has been merged gracefully.
+- __Note that__ duplicated features has been merged gracefully.
 
 ### Changed
 - sharp base version (from 0.23.3 to 0.23.4)